reshuffle api routes for scanning/update approvals
